Authentication-task
Refactor the project into a library and a binary: all the logic will live in the library crate while the
binary itself will be just an entrypoint with a very slim main function.

This is needed in order to be able to add a new tests folder for our integration tests.

Anything under tests folder ends up being compiled in its own binary - all our code under test is
imported as a crate. But the project, at the moment, is a binary: it is meant to be executed, not to be shared.
Therefore, we can’t import the main function in our tests as it is right now
